Stop exposing TmdbApi directly, implement needed methods in ApiHelper
Related to #6
