+++ app/src/main/java/com/arctouch/codechallenge/data/MovieDataSource.java
@@ -4,17 +4,8 @@ import android.arch.paging.PageKeyedDataSource;
 import android.support.annotation.NonNull;
 import android.util.Log;
 
-import com.arctouch.codechallenge.api.TmdbApi;
-import com.arctouch.codechallenge.model.Genre;
+import com.arctouch.codechallenge.api.ApiHelper;
 import com.arctouch.codechallenge.model.Movie;
-import com.arctouch.codechallenge.model.UpcomingMoviesResponse;
-import com.arctouch.codechallenge.util.ApiHelper;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-import retrofit2.Call;
 
 public class MovieDataSource extends PageKeyedDataSource<Long, Movie> {
     @SuppressWarnings("unchecked")
@@ -23,7 +14,7 @@ public class MovieDataSource extends PageKeyedDataSource<Long, Movie> {
         long page = (Long) params.key;
         Log.d("loadBefore", "called, page = " + page);
 
-        callback.onResult(loadPage(page), page - 1);
+        callback.onResult(ApiHelper.getUpcomingMovies(page), page - 1);
     }
 
     @SuppressWarnings("unchecked")
@@ -32,7 +23,7 @@ public class MovieDataSource extends PageKeyedDataSource<Long, Movie> {
         long page = (Long) params.key;
         Log.d("loadAfter", "called, page = " + page);
 
-        callback.onResult(loadPage(page), page + 1);
+        callback.onResult(ApiHelper.getUpcomingMovies(page), page + 1);
     }
 
     @SuppressWarnings("unchecked")
@@ -42,29 +33,6 @@ public class MovieDataSource extends PageKeyedDataSource<Long, Movie> {
 
         Cache.setGenres(ApiHelper.getGenres());
 
-        callback.onResult(loadPage(1), null, 2L);
-    }
-
-    private List<Movie> loadPage(long page) {
-        Call pageCall = ApiHelper.api.upcomingMoviesSynchronous(TmdbApi.API_KEY, TmdbApi.DEFAULT_LANGUAGE, page);
-
-        try {
-            UpcomingMoviesResponse response = (UpcomingMoviesResponse) pageCall.execute().body();
-
-            if (response != null) {
-                for (Movie movie : response.results) {
-                    movie.genres = new ArrayList<>();
-                    for (Genre genre : Cache.getGenres()) {
-                        if (movie.genreIds.contains(genre.id)) {
-                            movie.genres.add(genre);
-                        }
-                    }
-                }
-
-                return response.results;
-            }
-        } catch (Exception ignored) {}
-
-        return Collections.emptyList();
+        callback.onResult(ApiHelper.getUpcomingMovies(1), null, 2L);
     }
 }
