+++ app/src/main/java/com/arctouch/codechallenge/util/MovieImageUrlBuilder.java
@@ -1,6 +1,6 @@
 package com.arctouch.codechallenge.util;
 
-import com.arctouch.codechallenge.api.TmdbApi;
+import com.arctouch.codechallenge.api.ApiHelper;
 
 public class MovieImageUrlBuilder {
 
@@ -8,10 +8,10 @@ public class MovieImageUrlBuilder {
     private static final String POSTER_URL = "https://image.tmdb.org/t/p/w154";
 
     public String buildPosterUrl(String posterPath) {
-        return POSTER_URL + posterPath + "?api_key=" + TmdbApi.API_KEY;
+        return POSTER_URL + posterPath + "?api_key=" + ApiHelper.API_KEY;
     }
 
     public String buildBackdropUrl(String backdropPath) {
-        return BACKDROP_URL + backdropPath + "?api_key=" + TmdbApi.API_KEY;
+        return BACKDROP_URL + backdropPath + "?api_key=" + ApiHelper.API_KEY;
     }
 }
